requirements.txt
loguru==0.6.0
Pillow==9.4.0
playwright==1.30.0
